Extract ReflexInvoker
