+++ lib/stimulus_reflex.rb
@@ -12,6 +12,7 @@ require "stimulus_reflex/version"
 require "stimulus_reflex/reflex"
 require "stimulus_reflex/element"
 require "stimulus_reflex/channel"
+require "stimulus_reflex/service/reflex_invoker"
 require "stimulus_reflex/transport/base_adapter"
 require "stimulus_reflex/transport/cable_ready_adapter"
 require "generators/stimulus_reflex_generator"
